+++ packages/stryker-webpack-transpiler/src/compiler/Webpack.ts
@@ -1,0 +1,4 @@
+import * as webpack from 'webpack';
+
+// Used to mock away webpack during unit-testing
+export default webpack;