+++ packages/stryker-webpack-transpiler/src/index.ts
@@ -1,0 +1,4 @@
+import { TranspilerFactory } from 'stryker-api/transpile';
+import WebpackTranspiler from './WebpackTranspiler';
+
+TranspilerFactory.instance().register('webpack', WebpackTranspiler);